I changed to long

+++ 0x05-pointers_arrays_strings/6-puts2.c
@@ -8,7 +8,7 @@
 
 void puts2(char *str)
 {
-	int i = 0;
+	long i = 0;
 
 	for (i = 0; str[i]; i += 2)
 		_putchar(str[i]);
